+++ android/app/src/main/java/io/github/parkwithease/parkeasy/ui/history/bookings/BookingsNavigation.kt
@@ -1,4 +1,4 @@
-package io.github.parkwithease.parkeasy.ui.leasings
+package io.github.parkwithease.parkeasy.ui.history.bookings
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -6,16 +6,16 @@ import androidx.navigation.compose.composable
 import io.github.parkwithease.parkeasy.ui.common.enterAnimation
 import io.github.parkwithease.parkeasy.ui.common.exitAnimation
 
-private const val LeasingsRoute = "leasings"
+private const val BookingsRoute = "bookings"
 
-fun NavGraphBuilder.leasingsScreen() {
+fun NavGraphBuilder.bookingsScreen() {
     composable(
-        route = LeasingsRoute,
+        route = BookingsRoute,
         enterTransition = { enterAnimation() },
         exitTransition = { exitAnimation() },
     ) {
-        LeasingsScreen()
+        BookingsScreen()
     }
 }
 
-fun NavController.navigateToLeasings() = this.navigate(LeasingsRoute)
+fun NavController.navigateToBookings() = this.navigate(BookingsRoute)
